+++ src/client/handle_remote/socks5.rs
@@ -336,7 +336,7 @@ async fn handle_udp_relay_header<'buf>(
     // XXX: Note that we block on reading from the channel. This means that
     // only one client can use the channel at a time.
     let (len, addr) = socket.recv_from(buf).await?;
-    let _reserved = &buf[..2];
+    // let _reserved = &buf[..2];
     let frag = buf[2];
     if frag != 0 {
         warn!("Fragmented UDP packets are not implemented");
