+++ src/client/ws_connect.rs
@@ -2,7 +2,7 @@
 //! SPDX-License-Identifier: Apache-2.0 OR GPL-3.0-or-later
 
 use crate::arg::ServerUrl;
-use crate::mux::DEFAULT_WS_CONFIG;
+use crate::config;
 use crate::proto_version::PROTOCOL_VERSION;
 use crate::tls::make_rustls_client_config;
 use http::header::HeaderValue;
@@ -71,7 +71,8 @@ pub async fn handshake(
         Connector::Plain
     };
     let (ws_stream, _response) =
-        connect_async_tls_with_config(req, Some(DEFAULT_WS_CONFIG), Some(connector)).await?;
+        connect_async_tls_with_config(req, Some(config::DEFAULT_WS_CONFIG), Some(connector))
+            .await?;
     // We don't need to check the response now...
     debug!("WebSocket handshake succeeded");
     Ok(ws_stream)
